+++ src/core/index.ts
@@ -1,8 +1,8 @@
-export * from "./component.js";
 export * from "./context.js";
 export * from "./element-builder.js";
 export * from "./element-common.js";
 export * from "./env.js";
 export * from "./lifecycle.js";
 export * from "./signals.js";
+export type * from "./types.js";
 export * from "./view.js";