+++ src/core/view.ts
@@ -1,11 +1,10 @@
-import { Component } from "./component.js";
 import { NODE, NodeTarget } from "./element-common.js";
 import { ENV } from "./env.js";
-import { Falsy } from "./internals.js";
 import { createText } from "./internals/create-text.js";
 import { NOOP } from "./internals/noop.js";
 import { capture, nocapture, teardown, TeardownHook } from "./lifecycle.js";
 import { $, effect, Expression, ExpressionResult, get, memo, Signal, watch } from "./signals.js";
+import type { Component, Falsy } from "./types.js";
 
 /**
  * Internal utility to create placeholder comments.