+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 19.4
-+ Add support teardown hooks in `memo`.
++ Support teardown hooks in `memo`.
 
 ## 19.3
 + Add append fallback to `View.insertBefore`.